[Chore] Relocate components related to MyHistoryPage and TeamPage
